+++ xLights/effects/FireworksPanel.cpp
@@ -208,7 +208,7 @@ void FireworksPanel::SetTimingTracks(wxCommandEvent& event)
 
     // check if anything has been removed ... if it has clear the list and we will have to rebuild it as you cant delete items from a combo box
     bool removed = false;
-    for (int i = 0; i < Choice_TimingTrack->GetCount(); i++)
+    for (size_t i = 0; i < Choice_TimingTrack->GetCount(); i++)
     {
         bool found = false;
         for (auto it = timingtracks.begin(); it != timingtracks.end(); ++it)
